dp_rx/data/dp_rx.tcl: unbreak vid_phy direction
Commit edb1f978f719b (vid_phy_ctrl:hdmi_gt_ctrl: Generate Channels only if
connected to DP/HDMI) broke the vid_phy direction setting as it set the 4th cell to 1 instead of 0.

As specified by the binding, the 4th cell encodes the phy directions:

  direction_tx = 0 for RX and 1 for TX

https://github.com/Xilinx/dp-modules/blob/xlnx_rel_v2024.1/Documentation/devicetree/bindings/xlnx%2Cvphy.txt

So setting it to 1 doesn't make any sense for the DP receiver core.  This
seems to be a cut'n'paste mistake from the similar logic in the DP
transmitter core.

Fixes: edb1f978f719b ("vid_phy_ctrl:hdmi_gt_ctrl: Generate Channels only if
       connected to DP/HDMI")
Signed-off-by: Peter Korsgaard <peter.korsgaard@barco.com>